requirements.txt
# Django y utilidades
Django==5.0.1
dj-database-url==3.0.1
sqlparse==0.5.3
asgiref==3.10.0
tzdata==2025.2
whitenoise==6.11.0
gunicorn==23.0.0

# HTTP
requests==2.32.5
urllib3==2.5.0
idna==3.11
charset-normalizer==3.4.4
certifi==2025.10.5

# PostgreSQL
psycopg2-binary==2.9.11

# ML / Ciencia de datos (NECESARIO para cargar modelo .h5)
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2

# TensorFlow CPU (versión compatible, más ligera que la normal)
tensorflow-cpu==2.13.0
joblib==1.3.2
numpy==1.26.4

# Uvicorn (si lo usas local, no afecta)
uvicorn==0.37.0

# Compresión opcional
Brotli==1.1.0
colorama==0.4.6
click==8.3.0
packaging==25.0
